requirements-dev.txt
# additional requirements for running tests
pytest-mockito==0.0.4
testcontainers==4.9.1
